moved mc_facet_adjacency to new module
