+++ src/geometry/mod.rs
@@ -8,6 +8,5 @@ pub mod global_fcc_grid;
 pub mod grid_assignment_object;
 pub mod mc_cell_state;
 pub mod mc_domain;
-pub mod mc_facet_adjacency;
 pub mod mc_location;
 pub mod mesh_partition;
